Add utility for exploring `.safetensors` file
The interface provided by the official `safetensors` package is rather
clunky, and doesn't support any of the usual methods for interacting
with python objects.  (e.g. Using `st_file.get_tensor('name')` instead
of `st_file['name']`.)  In addition, not all fields from the metadata
can be displayed without either unpacking the metadata oneself, or
first reading a actual tensor from file.  (e.g. the shape/dtype of
tensors).

This is a quick and dirty way to read the safetensors files, inspect
the metadata (without reading the entire tensor!), and interact in a
more pythonic interface.