Pass `--flake path:.` instead of `--flake .`
Normally, submodules could be handled by setting
`inputs.self.submodules = true;`.  However, this can't check out a
private submodule, so it won't work in this case.  Passing `--flake
path:.` instead of `--flake .` bypasses the heuristcs that select the
flake type, treating it as a directory, and allow it to work correctly.